refactor(api): 调整 `SocialRelationsContainer` 及其子接口的包路径
