+++ simbot-apis/simbot-api/src/main/kotlin/love/forte/simbot/bot/Bot.kt
@@ -22,8 +22,7 @@ import kotlinx.coroutines.Job
 import kotlinx.coroutines.cancel
 import love.forte.simbot.*
 import love.forte.simbot.ability.Survivable
-import love.forte.simbot.definition.User
-import love.forte.simbot.definition.UserInfo
+import love.forte.simbot.definition.*
 import love.forte.simbot.event.EventProcessor
 import love.forte.simbot.message.Image
 import love.forte.simbot.message.Image.Key.toImage
