+++ simbot-apis/simbot-api/src/main/kotlin/love/forte/simbot/definition/SocialRelationsContainers.kt
@@ -14,7 +14,7 @@
  *
  */
 
-package love.forte.simbot.bot
+package love.forte.simbot.definition
 
 import love.forte.simbot.Api4J
 import love.forte.simbot.ID
